fix(tests): fix database setup for Django < 1.4
This fixes the TypeError that occurs during the database setup for the
testsuite in Django < 1.4. Removing some of the class attributes allows
for the 'inner()' method to be called from the BaseField 'instance'
instead of the PickedObjectField.

Example error:
TypeError: unbound method inner() must be called with BaseField instance
as first argument (got PickledObjectField instance instead)